[MINOR] Extracts KosherCertificateViewer component
This commit introduces a new component, `KosherCertificateViewer`, to handle the display of kosher certificates.

This change promotes reusability and improves code organization by extracting the certificate viewing logic from both the `KosherCertificatesCard` and `KosherCertificatesList` components into a single, dedicated component. This makes the codebase more maintainable and easier to understand.